+++ packages/cms/lib/modules/resource-form-widgets/index.js
@@ -134,7 +134,7 @@ module.exports = {
     self.load = function (req, widgets, next) {
       const styles = openstadMap.defaults.styles;
       const globalData = req.data.global;
-
+      
       widgets.forEach(async (widget) => {
         const resourceType = widget.resource ? widget.resource : false;
         const resourceInfo = resourceType ? resourcesSchema.find((resourceInfo) => resourceInfo.value === resourceType) : false;