+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+* Make image size configurable for the image widget
+* Fix HTML injection in search form in the resource overview widget
+* Add a new text style to select in the rich text editor.
+* Make autoCenter configurable in idea-map-widgets
+
 ## 1.0.0
 * Add submission to resource form with configurable confirmation settings
 * Update ideas-on-map config to work with react-leaflet